requirements.txt
# Core Data Science Libraries
pandas==2.0.3
numpy==1.24.3

# Machine Learning
scikit-learn==1.3.0

# Deep Learning (LSTM)
tensorflow-intel==2.13.0

# Model Serialization
joblib==1.3.2

# Visualization
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.16.1

# Web Application
streamlit==1.26.0

# Utilities
python-dateutil==2.8.2
